backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "speech-processing-backend"
version = "0.1.0"
description = "Speech-to-text processing backend with transcription, diarization, and translation"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    # Web Framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "websockets==12.0",

    # ML/AI - Core for MVP
    "faster-whisper==1.0.3",
    "torch==2.2.0",
    "torchaudio==2.2.0",

    # Audio Processing
    "librosa==0.10.2",
    "soundfile==0.13.1",
    "pydub==0.25.1",
    "av>=12.0.0",

    # Utilities
    "pydantic==2.8.2",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.0",
    "aiofiles==23.2.1",
    "numpy<2.0",

    # Storage
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",

    # Code Quality
    "black==24.10.0",
    "isort==5.12.0",
    "pylint==3.3.1",
    "mypy==1.13.0",
    "flake8==7.1.1",
]
diarization = [
    "pyannote.audio==3.1.0",
]
translation = [
    "transformers==4.35.0",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=html --cov-report=term"

[dependency-groups]
dev = [
    "black==24.10.0",
    "flake8==7.1.1",
    "httpx==0.28.1",
    "isort==5.12.0",
    "mypy==1.13.0",
    "pylint==3.3.1",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
]
